third_party/blink/web_tests/platform/linux-chrome/external/wpt/wai-aria/role/menu-roles-expected.txt

This is a testharness.js-based test.
[FAIL] orphaned menuitem outside the context of menu/menubar
  assert_equals: <nav role="menuitem" data-testname="orphaned menuitem outside the context of menu/menubar" data-expectedrole="navigation" class="ex">x\n</nav> expected "navigation" but got "menuitem"
[FAIL] orphaned menuitemradio outside the context of menu/menubar
  assert_equals: <nav role="menuitemradio" data-testname="orphaned menuitemradio outside the context of menu/menubar" data-expectedrole="navigation" class="ex">x\n</nav> expected "navigation" but got "menuitemradio"
[FAIL] orphaned menuitemcheckbox outside the context of menu/menubar
  assert_equals: <nav role="menuitemcheckbox" data-testname="orphaned menuitemcheckbox outside the context of menu/menubar" data-expectedrole="navigation" class="ex">x\n</nav> expected "navigation" but got "menuitemcheckbox"
[FAIL] orphan button with menuitem role
  assert_equals: <button role="menuitem" data-testname="orphan button with menuitem role" data-expectedrole="button" class="ex">x</button> expected "button" but got "menuitem"
[FAIL] orphan button with menuitemcheckbox role
  assert_equals: <button role="menuitemcheckbox" data-testname="orphan button with menuitemcheckbox role" data-expectedrole="button" class="ex">x</button> expected "button" but got "menuitemcheckbox"
[FAIL] orphan button with menuitemradio role
  assert_equals: <button role="menuitemradio" data-testname="orphan button with menuitemradio role" data-expectedrole="button" class="ex">x</button> expected "button" but got "menuitemradio"
Harness: the test ran to completion.
